Remove animated clouds
The animated clouds were a nice touch; however I decided to change up how
the front page will look. Because the front page is an attention-grabber,
a vertical scrolling one-page design seemed to be the best. But, I will
instead be moving to a design where the navigation sits in the
center and information flows from outside the viewport or from external
sources. This new look requires plenty of refactoring.

- Remove all instances of animated clouds from index.html
- Remove all styles associated with animated clouds
- Rename .cloud-icon to .cloud in navigation (undo change made in
  commit 7868c3e)