Add parameter for stylesheet to use
Jekyll allows passing of variables when including assets. In the case of
layouts, developers can pass parameters that can be plugged into the
included files as {{ include.param_name }}. A css parameter is now passed
from the layouts to head.html indicating whether splash.css or main.css is
loaded on the page. If it is the home page, then a string "splash" is
passed; otherwise, "main" is passed.