Remove height from right section of navigation
The height was reinforced in an earlier build of the site to work with
mobile platforms of various sizes. However, the optimization for mobile
devices as of recently no longer requires this hackish fix, and leaving it
as is causes the right section to stand a layer above the left section.
The title area ends up being covered up and the link to the front page
cannot be clicked. Removing the height resolves this issue.