Add smart pagination
Pagination was set to print all pages on the screen. This will be
unmanageable in the future if there are several pages of posts, as it will
fill up the viewport. Ideally, a fixed number of pages would be desired on
the viewport. By using the site.paginate variable from the config file,
you can fix the pagination such that it will not allow you to cross past
the midpoint. In other words, if you were to pass page 3 and are only
showing 5 pages on the viewport, rather than going on to page 4, the 5
pages on the screen will shift such that 4 is now in the center. For the
current test case, there are 9 pages. 5 pages show up at most at any time.
Page 4 causes a shift so that you see pages 2-6. The number of pages on
the viewport is still fixed to 5. Moving on to page 5 shifts the pages on
the screen to 3-7. On page 7, it is shifted to 5-9. Page 8 and 9 will pass
the midpoint and maintain the view of pages 5-9. This scales to any number
of pages, keeping the fixed amount of 5.

The fixed amount is derived from the site.paginate variable - you can
never be more than site.paginate pages away from the "previous" or "next"
labels. In the case of site.paginate = 3, that 3 from the left + 3 from
the right. Subtract 1 for excluding the middle and you get 5 as the fixed
number of pages shown on the viewport.